Rename cs_VERSION to controlfilter_VERSION.
The name of this variable needs to match the name used for the submodule
in x42-plugins, in order for the rule that generates gitversion.mak to
work correctly. If the name doesn't match, building x42-plugins from a
tarball will result in it trying to figure out the version based on the
Git tag and failing.